WIP refactor for v20+
